feat(gui): change class structure to support retained mode rendering
Immediate mode is considered an old way of rendering in OpenGL, and
although it is still supported it is far less efficient than retained
mode. This is a more modern way of using OpenGL that stores objects in
vertex buffers and allows the actual rendering process to be optimized
with batching/culling/etc. at draw time rather than calling a function
for each primitive. This pipeline also better matches the idea of
rendering many cells at once--by treating each cell as a vertex rather
than a collection of primitives--and allows all cells to be drawn
simultaneously instead of in a loop.

BREAKING CHANGE: The new class requires Simulator to expose renderers
that create and manage vertex buffers and provide draw calls. GUI
functionality is broken as of this commit until the rest of the classes
are implemented.