feat(simulation): add abstract base class for flexible module types
In order to better support the ability to easily implement and clearly
specify additional functionality to CellModeller, there should be a
strong but flexible specification for what methods are required to be
included in modules and can be integrated into the main Simulator. The
ModuleProtocol class is one such solution that abstracts module
behaviour away from integration with the simulation and managing cell
attributes. It also includes a specification for triggering and
handling cell events across modules, as well as an extension for user-
defined simulation configurations via UserModule.