feat(biophysics): add default biophysics class for bacillus cells
This class is a concrete implementation of CMModule that abstracts the
core mathematical logic of the simulation into modular and specific
functions that can be reused or reimplemented by any biophysics class
deriving from the basic bacterial model. It includes the addition of
Contact objects which describe various physical interactions that can
occur in a simulation. As a consequence of modularizing the biophysics
handling and moving away from a monolithic kernel, CellModeller is now
able to integrate better with existing scientific computing libraries
for Python, such as SciPy, which includes a validated implementation of
an efficient sparse matrix iterative solver in scipy.sparse.linalg.lsmr
Using scipy sparse matrix objects also has the benefit of making
equations involving matrix operations explicit and easy to read.